+++ src/utils.c
@@ -6,9 +6,9 @@
 /**
  * Affiche un tableau de double
  */
-void afficher_tableau(double *array, size_t n) {
+void afficher_tableau(long double *array, size_t n) {
     for (size_t i = 0; i < n; ++i) {
-        printf("%lf - ", array[i]);
+        printf("%Lf - ", array[i]);
     }
     printf("\n");
 }
@@ -19,11 +19,11 @@ void afficher_tableau(double *array, size_t n) {
  * @param array tableau de taille (n * (n + 1)) / 2
  * @param n taille de la matrice carrée
  */
-void afficher_tableau_trig_double(double *array, size_t n) {
+void afficher_tableau_trig_double(long double *array, size_t n) {
     size_t line = 1;
     size_t col = 0;
     for (size_t i = 0; i < (n * (n + 1)) / 2; ++i) {
-        printf("%lf - ", array[i]);
+        printf("%Lf - ", array[i]);
         if (col == n - line) {
             ++line;
             col = 0;
